fix: rename title to targetTable
